Add async generator API to pywatchman_aio
Summary:
Many of the other client libraries have event driven async APIs, but python does not really have that currently. The best it has is `get_subscription` which acts as a polling api as it directly returns `q.get()` instead of returning the queue itself.

This diff adds an API to yield elements in the queue so that it can be used as an async generator to make the code more event driven.

Test Plan:
tbd